requirements.txt
disnake
pyyaml
python-dotenv